Move context storage property set to later, add more tests
